domain/core-domain/src/main/kotlin/com/example/coredomain/voucher/model/Voucher.kt
package com.example.coredomain.voucher.model

import com.example.coredomain.common.type.VoucherStatus
import com.example.coredomain.contract.model.Contract
import com.example.coredomain.voucherproduct.model.VoucherProduct
import java.time.LocalDate
import java.time.LocalDateTime
import java.util.*

data class Voucher(
    val code: String = UUID.randomUUID().toString().uppercase().replace("-", ""),
    var expiredDate: LocalDate? = null,
    val description: String? = null,

    val contract: Contract,
    val product: VoucherProduct,
    val histories: MutableList<VoucherHistory> = mutableListOf(
        VoucherHistory(
            voucherCode = code,
            voucherStatus = VoucherStatus.ISSUED
        )
    ),
) {
    val status: VoucherStatus
        get() = histories.last().voucherStatus

    val issuedAt: LocalDateTime
        get() = histories.first { it.voucherStatus == VoucherStatus.ISSUED }.createAt
}
